Switch sign convention for Zeeman term (to the actual one) (#12)
* Switch sign convention for Zeeman term (to the actual one)

In previous script versions, the magnetic moments due to the pseudo-spin
were missing a sign, which was compensated by the missing sign in the
previous definition of the Zeeman term. But actually:

  M = - g S   (where S is a pseudospin operator, M and S are vectorial) and
  H_Zeeman = - M.B     (B is the magn. field and "." the scalar product)

Scripts using diagonalizeSpinHamiltonian() must be updated!

* Switch sign convention for Zeeman term (to the actual one)

In previous script versions, the magnetic moments due to the pseudo-spin
were missing a sign, which was compensated by the missing sign in the
previous definition of the Zeeman term. But actually:

  M = - g S   (where S is a pseudospin operator, M and S are vectorial) and
  H_Zeeman = - M.B     (B is the magn. field and "." the scalar product)

Scripts using diagonalizeSpinHamiltonian() must be updated!